Add GetFileName() method to MyCurl struct
The GetFileName() method has been added to the MyCurl struct in the GmailAPI.h file. This method extracts the file name from a path stored in the result string by iterating backward from the end of the string until it encounters a '/' character. It then reverses the extracted file name and returns it.